@ v2.3: 游戏机制 + 奖惩函数 + 训练稳定性 全面升级
### 游戏机制 (Unity C#)
- 球碰墙反弹替代重置 (0.92系数)，球不再消失
- 四角惩罚 (-0.15)，对方微奖励 (+0.05)
- 球门物理建模: 仅正面进球，侧面/顶部/后面反弹
- GoalTrigger 精准匹配 Soccergoal 模型尺寸
- 清理冗余 GoalNet 标签 + 废弃 FieldGoals_Root
- BlueStriker isHumanControlled=true, P1/P2 键位
- 训练时全部 Agent BehaviorType=Default

### Python 训练端
- 新增奖励塑形模块 (reward_shaper.py): Potential Shaping + Curriculum + Normalize
- 训练补丁从2项扩展到6项: 梯度裁剪 + lambda_return修复 + 优势归一化 + 熵衰减 + 价值裁剪
- lambda_return object array 修复 (进球后轨迹不等长bug)
- redis_writer numpy.float32 JSON序列化修复
- config/poca_soccer.yaml 优化 + 新增 reward_shaping.yaml
- train.py v2.3: 自动加载补丁 + 奖励塑形器

### 文档
- README.md v2.3 更新日志

@

+++ python/rlsoccer/redis_writer.py
@@ -145,10 +145,10 @@ class RedisStatsWriter(StatsWriter):
         self._step_count += 1
         timestamp = time.time()
 
-        # 提取关键指标
+        # 提取关键指标 (确保转为 Python 原生 float)
         flat: Dict[str, float] = {}
         for key, summary in values.items():
-            flat[key] = round(summary.mean, 6)
+            flat[key] = float(round(summary.mean, 6))
 
         # ── 按指标类型写入不同 Redis Key ──
